requirements.txt
cliff
imapclient
lxml
PyYAML
pbr
